Change build process
Shared media assets are not copies to destination but to the source. This is necessary for Jekyll static files list and it will be necessary for implicit thumbnails in the future.

File cloning on macOS must be explicit (unlike on Linux). However, it is explicit on LInux now, so that the use is warned that large number of files are copies around if cloning fails.